server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
supabase 